feat(auth): 增加对 Google、GitHub 和 Apple 的第三方登录支持
在用户登录和注册的请求结构中新增 Google、GitHub 和 Apple 登录选项，更新相关数据模型以支持这些新登录方式。确保系统能够处理多种第三方登录，增强用户体验。

+++ service/syssvr/internal/logic/tenantmanage/tenantAppUpdateLogic.go
@@ -56,7 +56,16 @@ func (l *TenantAppUpdateLogic) TenantAppUpdate(in *sys.TenantAppInfo) (*sys.Empt
 	if in.Huawei != nil {
 		old.Huawei = utils.Copy[relationDB.SysTenantThird](in.Huawei)
 	}
-	if in.Android.IsUpdateFilePath {
+	if in.Google != nil {
+		old.Google = utils.Copy[relationDB.SysTenantThird](in.Google)
+	}
+	if in.Github != nil {
+		old.Github = utils.Copy[relationDB.SysTenantThird](in.Github)
+	}
+	if in.Apple != nil {
+		old.Apple = utils.Copy[relationDB.SysTenantAppleConfig](in.Apple)
+	}
+	if in.Android != nil && in.Android.IsUpdateFilePath {
 		if old.Android != nil && old.Android.FilePath != "" {
 			err := l.svcCtx.OssClient.PublicBucket().Delete(l.ctx, old.Android.FilePath, common.OptionKv{})
 			if err != nil {
